pyproject.toml
[project]
name = "py2glsl"
dynamic = ["version"]
description = "Project description"
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "glfw>=2.8.0",
    "imageio[ffmpeg]>=2.37.0",
    "loguru>=0.7.3",
    "moderngl>=5.12.0",
    "numpy>=2.2.2",
    "pillow>=11.1.0",
]

[project.urls]
homepage = "https://github.com/alexeykarnachev/py2glsl"
repository = "https://github.com/alexeykarnachev/py2glsl"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "py2glsl/__init__.py"

[tool.hatch.build]
include = [
    "py2glsl/**/*.py",
    "py2glsl/**/*.md",
]

[tool.hatch.build.targets.wheel]
packages = ["py2glsl"]

[project.scripts]
py2glsl = "py2glsl.main:app"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "asyncio: mark test as async/await test",
]

[tool.ruff]
target-version = "py313"
line-length = 88
indent-width = 4
extend-exclude = [".pytest_cache", ".ruff_cache", ".venv", "venv"]

[tool.ruff.lint]
select = [
    "F",     # Pyflakes
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "I",     # isort
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "A",     # flake8-builtins
    "C4",    # flake8-comprehensions
    "SIM",   # flake8-simplify
    "ERA",   # eradicate
    "PL",    # pylint
    "RUF",   # ruff-specific rules
    "F841",  # unused variables
]
per-file-ignores = { "__init__.py" = ["F401"], "tests/*" = ["PLR2004"] }

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true

[dependency-groups]
dev = [
    "mypy>=1.9.0",
    "pytest>=8.1.1",
    "pytest-asyncio>=0.23.5.post1",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.3",
    "types-pyyaml>=6.0.12.20240311",
]
